Add fourth var to line shape model to follow 10.1016/j.jqsrt.2019.106676
This removes the old interpolation variable so IO will not work with these
files anymore.  The general workpace to regenerate the catalogs still work
as before so I find it acceptable to have the last two months of stored
catalogs not work.  If it is not acceptable, please let me know