Add abs_lines_per_species2 and a way to create it from abs_lines2
There is a small design detail difference between this version and
for the version in abs_lines_per_species.  It used to be possible
to have a one-to-one match between abs_species and
abs_lines_per_species.  However, since the subdivision is so much
greater potentiall in the new version of the catalog, there is no
way to keep the inner layer the same without introducing yet
another "ArrayOf" outside.  Is this a problem?  I don't think it
is.  At least not for any of the forward calculations.