xsec forward calculations now works
Add temporarily abs_xsec_per_speciesAddLines3 to keep around
while debugging.  The goal is to make it abs_xsec_per_speciesAddLines
and remove the other two before this is merged

The code agrees perfectly with abs_xsec_per_speciesAddLines2
but there is an offset towards abs_xsec_per_speciesAddLines.

I have yet to investigate any jacobian or normalization or nlte or
cutoff or mirroring setups.  So this code is not production ready

Also, the NLTE code is broken in abs_xsec_per_speciesAddLines3.
I need a new matching-mechanism.  There is the beginnings of one
one in quantum.h/cc but this is also far from ready.  Especially,
it is a compound class that needs IO to work properly.  Also, it
needs to exist as a copy, since I need one to replace the current
nlte_field (Tensor3-level) and one to replace the current abs_nlte
(Vector-level).  So IO will be annoying to deal with...