Add ugly workaround for NLTE vibrational temperatures
Note, I spotted a bug!  If S_src and S_abs are swapped then the old code
and new code will agree much more closely for vib-temp NLTE.  However, I
have been unable to locate the exact theoretical happening where upper
and lower states are swapped, so I have no solution to add at this point.

Note, there is a lot of *2 functions and agendas in this branch now.
There is even a *3...  these will be removed before merging the branch
with mainline...